backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
idna==3.6
MarkupSafe==2.1.3
mongoengine==0.27.0
PyJWT==2.8.0
pymongo==4.6.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
